+++ NEWS.md
@@ -78,6 +78,8 @@
 * Now one-line empty objects, e.g. `"Output:Surfaces:List,,;\n"`, can be
   successfully parsed (#88).
 * Fix the error of `$to_table()` when no arguments were provided (#103).
+* The references of newly added extensible fields are now correctly handled
+  (#109).
 
 ## Minor changes